Major performance improvements, OTFFT support
I switched to a new method of performing Chebyshev expansion, which doesn't require any padding of the input array.
I also added the option to compile with OTFFT support, which is much faster than KissFFT which is the default, and is the best options I could find for my use case.
Spectra is now an optional dependency to shorten compile times.
When OTFFT is used, the FFT is no longer the bottleneck for the calculation. Because of this, I also use an approximation for calculating the Chebyshev nodes by using the angle additional formulae instead of directly calling cos(theta) each time. This slightly affects accuracy but for my case the speed is worth it, I will make it optional in future.