New features, bug fixes.
I can't remember everything I changed but from memory I reworked the functions slightly so that I could create a few variants of each. Now it's possible to get a Chebyshev expansion to a given error tolerance rather than to a given number of terms, and the same for CF approximation.